First steps
Remove process actions

Remove custom payment status from changelist

Make Registrations and Renewals Payable

Make payment widget a todo for Sebas

Fix the payment widget

Fix create_payments

Fix things

Add mandate for registrations and change process signal

Add manual migration

translations

Fix

Remove payment None and payment processed

+++ website/payments/services.py
@@ -41,7 +41,6 @@ def create_payment(
             notes=payable.payment_notes,
             topic=payable.payment_topic,
             paid_by=payable.payment_payer,
-            processing_date=timezone.now(),
             type=pay_type,
         )
     return payable.payment
@@ -54,43 +53,12 @@ def delete_payment(payable: Payable):
     :return:
     """
     payment = payable.payment
+    if payment.created_at < datetime.datetime.now() - datetime.timedelta(minutes=10):
+        raise PermissionError(_("You are not authorized to delete this payment."))
+
     payable.payment = None
     payable.save()
     payment.delete()
-
-
-def process_payment(
-    queryset: QuerySet, processed_by: Member, pay_type: str = Payment.CARD
-) -> list:
-    """
-    Process the payment
-
-    :param queryset: Queryset of payments that should be processed
-    :type queryset: QuerySet[Payment]
-    :param processed_by: Member that processed this payment
-    :type processed_by: Member
-    :param pay_type: Type of the payment
-    :type pay_type: String
-    """
-
-    queryset = queryset.filter(type=Payment.NONE)
-    data = []
-
-    # This should trigger post_save signals, thus a queryset update
-    # is not appropriate, moreover save() automatically sets
-    # the processing date
-    for payment in queryset:
-        if pay_type != Payment.TPAY or (
-            pay_type == Payment.TPAY
-            and payment.paid_by
-            and payment.paid_by.tpay_enabled
-        ):
-            payment.type = pay_type
-            payment.processed_by = processed_by
-            payment.save()
-            data.append(payment)
-
-    return data
 
 
 def update_last_used(queryset: QuerySet, date: datetime.date = None) -> int:
